+++ app/src/main/java/com/challenge/weather/presentation/features/weather/model/WeatherModel.kt
@@ -7,7 +7,7 @@ import kotlinx.parcelize.Parcelize
 data class WeatherModel(
     val country: String,
     val city: String,
-    val date: Double,
+    val date: Long,
     val icon: String,
     val description: String,
     val temperature: Double,
